requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cssselect==1.1.0
fake-useragent==0.1.11
idna==3.2
importlib-metadata==4.8.1
lxml==4.6.3
parse==1.19.0
pyee==8.2.2
pyppeteer==0.2.6
pyquery==1.4.3
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
requests-html==0.10.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tqdm==4.62.2
typing-extensions==3.10.0.2
tzlocal==3.0
urllib3==1.26.6
w3lib==1.22.0
websockets==9.1
zipp==3.5.0
